Fix build for python 3.9 and up
The build wasn't working for python 3.9 because `tp_print` was
deprecated but it was trying to use the already generated dubins.c file.
To fix this, I added a pyproject.toml file that guarantees that Cython
is available and I switched to using the `cythonize` call in order to
skip using a pre-generated `dubins/dubins.c` file.